% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interceptation.R
\name{CInt_f}
\alias{CInt_f}
\title{Canopy interceptation}
\usage{
CInt_f(R, Rstar)
}
\arguments{
\item{R}{Rainfall}

\item{Rstar}{The maximum amount which the canopy intercepts}
}
\value{
canopy interceptation
}
\description{
This function calculates the amount of rain intercepted in the canopy.
}
\details{
Interceptation
}
\examples{
CInt_f(R = 10, Rstar = 3)
}
