% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leakage.R
\name{Lk_f}
\alias{Lk_f}
\title{Leakage}
\usage{
Lk_f(s, Ks, b)
}
\arguments{
\item{s}{Soil moisture}

\item{Ks}{Soil saturated hydraulic conductivity}

\item{b}{The exponent of the water retention curve}
}
\value{
leakage
}
\description{
This function calculates the leakage based on the soil moisture, soil water retantion curve and the soil hydraulic conductivity.
}
\details{
Leakage function based on the soil moisture
}
\examples{
Lk_f(s = 0.25, Ks = 2000, b = 4.38)
}
