% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rainfall.R
\name{RainPoisson}
\alias{RainPoisson}
\title{Rainfall series}
\usage{
RainPoisson(ndays, lambda, alpha)
}
\arguments{
\item{ndays}{Number of days}

\item{lambda}{The frequency of rainfall events (day^-1)}

\item{alpha}{The mean of rainfall event (cm day^-1)}
}
\value{
rainfall series
}
\description{
This function simulates rainfall series as a stochastic variable, by using marked Poisson process.
}
\details{
Rainfall series
}
\examples{
RainPoisson(ndays = 60, lambda = 0.1, alpha = 0.95)
}
